// src/app/lib/supabase.types.ts
export interface Database {
  public: {
    Tables: {
      items: {
        Row: {
          created_at: string | null
          description: string
          id: string
          images: string[]
          max_stock: number
          min_stock: number
          payment_link: string | null
          price: number
          title: string
        }
        Insert: {
          created_at?: string | null
          description: string
          id?: string
          images?: string[]
          max_stock?: number
          min_stock?: number
          payment_link?: string | null
          price: number
          title: string
        }
        Update: {
          created_at?: string | null
          description?: string
          id?: string
          images?: string[]
          max_stock?: number
          min_stock?: number
          payment_link?: string | null
          price?: number
          title?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}